24-update

+++ src/addresses/addresses.resolver.ts
@@ -4,9 +4,10 @@ import { Address } from './entities/address.entity';
 import { CreateAddressInput } from './dto/create-address.input';
 import { UpdateAddressInput } from './dto/update-address.input';
 import { AccessAuthGuard } from 'src/guards/accessAuth.guard';
-import { UseGuards } from '@nestjs/common';
+import { BadRequestException, UseGuards } from '@nestjs/common';
 import { CheckAbilities } from 'src/ability/ability.decorator';
 import { Actions } from 'src/ability/ability.factory';
+import { isValidObjectId } from 'mongoose';
 
 @UseGuards(AccessAuthGuard)
 @Resolver(() => Address)
@@ -14,32 +15,36 @@ export class AddressesResolver {
   constructor(private readonly addressesService: AddressesService) {}
 
   @Mutation(() => Address)
-  @CheckAbilities({actions: Actions.Create, subject: Address})
+  @CheckAbilities({actions: Actions.Create, subject: "User"})
   createAddress(@Args('createAddressInput') createAddressInput: CreateAddressInput) {
     return this.addressesService.create(createAddressInput);
   }
 
   @Query(() => [Address], { name: 'addresses' })
-  @CheckAbilities({actions: Actions.Read, subject: Address})
+  @CheckAbilities({actions: Actions.Read, subject: "User"})
   findAll(@Args('id', {type: () => ID}) id: string) {
+    if(!isValidObjectId(id)) throw new BadRequestException("There isn't address with this id");
     return this.addressesService.findAll(id);
   }
 
   @Query(() => Address, { name: 'address' })
-  @CheckAbilities({actions: Actions.Read, subject: Address})
+  @CheckAbilities({actions: Actions.Read, subject: "User"})
   findOne(@Args('id', { type: () => ID }) id: string) {
+    if(!isValidObjectId(id)) throw new BadRequestException("There isn't address with this id");
     return this.addressesService.findOne(id);
   }
 
   @Mutation(() => String)
-  @CheckAbilities({actions: Actions.Update, subject: Address})
+  @CheckAbilities({actions: Actions.Update, subject: "User"})
   updateAddress(@Args('updateAddressInput') updateAddressInput: UpdateAddressInput) {
+    if(!isValidObjectId(updateAddressInput?.id)) throw new BadRequestException("There isn't address with this id");
     return this.addressesService.update(updateAddressInput.id, updateAddressInput);
   }
 
   @Mutation(() => String)
-  @CheckAbilities({actions: Actions.Delete, subject: Address})
+  @CheckAbilities({actions: Actions.Delete, subject: "User"})
   removeAddress(@Args('id', { type: () => ID }) id: string) {
+    if(!isValidObjectId(id)) throw new BadRequestException("There isn't address with this id");
     return this.addressesService.remove(id);
   }
 }
