+++ src/tags/tags.resolver.ts
@@ -4,11 +4,12 @@ import { Tag } from './entities/tag.entity';
 import { CreateTagInput } from './dto/create-tag.input';
 import { UpdateTagInput } from './dto/update-tag.input';
 import { AccessAuthGuard } from 'src/guards/accessAuth.guard';
-import { UseGuards } from '@nestjs/common';
+import { BadRequestException, UseGuards } from '@nestjs/common';
 import { CheckAbilities } from 'src/ability/ability.decorator';
 import { Actions } from 'src/ability/ability.factory';
 import { StateInput } from 'src/constants/state.input';
 import { UpdatePositionInput } from 'src/constants/position.input';
+import { isValidObjectId } from 'mongoose';
 
 @UseGuards(AccessAuthGuard)
 @Resolver(() => Tag)
@@ -16,44 +17,48 @@ export class TagsResolver {
   constructor(private readonly tagsService: TagsService) {}
 
   @Mutation(() => Tag)
-  @CheckAbilities({actions: Actions.Create, subject: Tag})
+  @CheckAbilities({actions: Actions.Create, subject: "Tag"})
   createTag(@Args('createTagInput') createTagInput: CreateTagInput) {
     return this.tagsService.create(createTagInput, null);
   }
 
   @Query(() => [Tag], { name: 'tags' })
-  @CheckAbilities({actions: Actions.Read, subject: Tag})
+  @CheckAbilities({actions: Actions.Read, subject: "Tag"})
   findAll() {
     return this.tagsService.findAll();
   }
 
   @Query(() => Tag, { name: 'tag' })
-  @CheckAbilities({actions: Actions.Read, subject: Tag})
+  @CheckAbilities({actions: Actions.Read, subject: "Tag"})
   findOne(@Args('id', { type: () => ID }) id: string) {
+    if(!isValidObjectId(id)) throw new BadRequestException("There isn't tag with this id");
     return this.tagsService.findOne(id);
   }
 
   @Mutation(() => String)
-  @CheckAbilities({actions: Actions.Update, subject: Tag})
+  @CheckAbilities({actions: Actions.Update, subject: "Tag"})
   updateTag(@Args('updateTagInput') updateTagInput: UpdateTagInput) {
+    if(!isValidObjectId(updateTagInput?.id)) throw new BadRequestException("There isn't tag with this id");
     return this.tagsService.update(updateTagInput.id, updateTagInput);
   }
 
   @Mutation(() => String)
-  @CheckAbilities({actions: Actions.Update, subject: Tag})
+  @CheckAbilities({actions: Actions.Update, subject: "Tag"})
   stateTag(@Args('stateInput') stateInput: StateInput) {
+    if(!isValidObjectId(stateInput?.id)) throw new BadRequestException("There isn't tag with this id");
     return this.tagsService.state(stateInput);
   }
 
   @Mutation(() => String)
-  @CheckAbilities({actions: Actions.Update, subject: Tag})
+  @CheckAbilities({actions: Actions.Update, subject: "Tag"})
   positionTag(@Args('updatePositionInput', {type: ()=> [UpdatePositionInput]}) updatePositionInput: UpdatePositionInput[]) {
     return this.tagsService.position(updatePositionInput);
   }
 
   @Mutation(() => String)
-  @CheckAbilities({actions: Actions.Delete, subject: Tag})
+  @CheckAbilities({actions: Actions.Delete, subject: "Tag"})
   removeTag(@Args('id', { type: () => ID }) id: string) {
+    if(!isValidObjectId(id)) throw new BadRequestException("There isn't tag with this id");
     return this.tagsService.remove(id);
   }
 }
